% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-dev-versions.R
\name{install_dev_versions}
\alias{install_dev_versions}
\title{Installs all development versions of RECON packages from Github}
\usage{
install_dev_versions()
}
\description{
Be careful this command may overwrite your existing CRAN packages.
}
